+++ src/org/joml/Matrix4x3dc.java
@@ -82,7 +82,7 @@ public interface Matrix4x3dc {
     /**
      * @return the properties of the matrix
      */
-    byte properties();
+    int properties();
 
     /**
      * Return the value of the matrix element at column 0 and row 0.
