+++ src/org/joml/Matrix4x3f.java
@@ -57,7 +57,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
     float m20, m21, m22;
     float m30, m31, m32;
 
-    byte properties;
+    int properties;
 
     /**
      * Create a new {@link Matrix4x3f} and set it to {@link #identity() identity}.
@@ -199,7 +199,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
     /* (non-Javadoc)
      * @see org.joml.Matrix4x3fc#properties()
      */
-    public byte properties() {
+    public int properties() {
         return properties;
     }
 
@@ -520,7 +520,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         } else {
             setMatrix4fc(m);
         }
-        properties = (byte) (m.properties() & (PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        properties = m.properties() & (PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
         return this;
     }
     private void setMatrix4fc(Matrix4fc mat) {
@@ -840,7 +840,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = nm30;
         dest.m31 = nm31;
         dest.m32 = nm32;
-        dest.properties = (byte) (this.properties & right.properties() & PROPERTY_ORTHONORMAL);
+        dest.properties = this.properties & right.properties() & PROPERTY_ORTHONORMAL;
         return dest;
     }
 
@@ -872,7 +872,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = nm30;
         dest.m31 = nm31;
         dest.m32 = nm32;
-        dest.properties = (byte) (right.properties() & PROPERTY_ORTHONORMAL);
+        dest.properties = right.properties() & PROPERTY_ORTHONORMAL;
         return dest;
     }
 
@@ -3008,7 +3008,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
         return dest;
     }
 
@@ -3063,7 +3063,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = nm30;
         dest.m31 = nm31;
         dest.m32 = nm32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
         return dest;
     }
 
@@ -3132,7 +3132,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3202,7 +3202,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3272,7 +3272,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3390,7 +3390,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3486,7 +3486,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3582,7 +3582,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3661,7 +3661,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3773,7 +3773,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
 
         return dest;
     }
@@ -3852,7 +3852,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = nm30;
         dest.m31 = nm31;
         dest.m32 = nm32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -3972,7 +3972,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m00 * x + m10 * y + m20 * z + m30;
         dest.m31 = m01 * x + m11 * y + m21 * z + m31;
         dest.m32 = m02 * x + m12 * y + m22 * z + m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY));
+        dest.properties = properties & ~(PROPERTY_IDENTITY);
         return dest;
     }
 
@@ -4092,7 +4092,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30 + x;
         dest.m31 = m31 + y;
         dest.m32 = m32 + z;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY));
+        dest.properties = properties & ~(PROPERTY_IDENTITY);
         return dest;
     }
 
@@ -4212,7 +4212,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = m20 * rm22;
         dest.m21 = m21 * rm22;
         dest.m22 = m22 * rm22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -4381,7 +4381,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = m20 * rm22;
         dest.m21 = m21 * rm22;
         dest.m22 = m22 * rm22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -4686,7 +4686,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = m20 * rm22;
         dest.m21 = m21 * rm22;
         dest.m22 = m22 * rm22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -4850,7 +4850,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = m20 * rm22;
         dest.m21 = m21 * rm22;
         dest.m22 = m22 * rm22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -5140,7 +5140,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = -m20;
         dest.m21 = -m21;
         dest.m22 = -m22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -5230,7 +5230,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = m20;
         dest.m21 = m21;
         dest.m22 = m22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -5490,7 +5490,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
 
         return dest;
     }
@@ -5886,7 +5886,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m10 = nm10;
         dest.m11 = nm11;
         dest.m12 = nm12;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -6194,7 +6194,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m10 = nm10;
         dest.m11 = nm11;
         dest.m12 = nm12;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
 
         return dest;
     }
@@ -6311,7 +6311,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -6408,7 +6408,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = m30;
         dest.m31 = m31;
         dest.m32 = m32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -6482,7 +6482,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m30 = nm30;
         dest.m31 = nm31;
         dest.m32 = nm32;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -6677,7 +6677,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m10 = nm10;
         dest.m11 = nm11;
         dest.m12 = nm12;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
 
         return dest;
     }
@@ -7437,7 +7437,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m20 = nm20;
         dest.m21 = nm21;
         dest.m22 = nm22;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION | PROPERTY_ORTHONORMAL);
 
         return dest;
     }
@@ -7785,7 +7785,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
      */
     public Matrix4x3f swap(Matrix4x3f other) {
         MemUtil.INSTANCE.swap(this, other);
-        byte props = properties;
+        int props = properties;
         this.properties = other.properties;
         other.properties = props;
         return this;
@@ -7826,7 +7826,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m00 = nm00;
         dest.m01 = nm01;
         dest.m02 = nm02;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
 
@@ -8181,7 +8181,7 @@ public class Matrix4x3f implements Externalizable, Matrix4x3fc {
         dest.m10 = nm10;
         dest.m11 = nm11;
         dest.m12 = nm12;
-        dest.properties = (byte) (properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION));
+        dest.properties = properties & ~(PROPERTY_IDENTITY | PROPERTY_TRANSLATION);
         return dest;
     }
