+++ src/org/joml/Matrix4x3fc.java
@@ -85,7 +85,7 @@ public interface Matrix4x3fc {
     /**
      * @return the properties of the matrix
      */
-    byte properties();
+    int properties();
 
     /**
      * Return the value of the matrix element at column 0 and row 0.
